backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
supabase==2.5.0
python-dotenv==1.0.1
pydantic==2.7.1
groq==0.9.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pymysql==1.1.1
cryptography==42.0.8
httpx==0.27.0
